Implement saving of PaperMC server definitions
This commit adds the ability to save PaperMC server definitions in the database. A new function, save_paper_mc_server, was introduced in the database module to handle this task. It inserts a new record for each of Application, MinecraftServer, and PaperMcServer entities. It also updates the create_paper_mc_server method in the IronMinecraftServerCreator struct of the web module to call the new method and return a newly created server id. The ServerDefinition message in the server.proto file was slightly modified - changing 'Name' field to 'name'.