Add ApplicationManager to handle server instances
Integrate an `ApplicationManager` object into the `IronMinecraftServerCreator` structure to manage running server instances. This `ApplicationManager` starts a new server after each successful database entry. The `save_paper_mc_server` function was adjusted to return an `Application` object that represents the newly saved server. This allows for the application to be directly passed into the `ApplicationManager` for instant startup.

In addition, a new `manager` module was created to organize associated structures and logic. Changes to some functions and data structures were necessary to accommodate this updated server management process.