"Refactor application's data structure and improve grpc service
The current approach of handling applications and their data had become confusing and was hard to maintain. The data flow was changed in order to simplify and unify the process - the refactor involves entity mapping in database module, web module's grpc service and the application manager. The data structure for Applications has been reworked to clearly distinguish between active and inactive Applications and their handling methods have been updated accordingly.

To simplify the process of saving PaperMcServers, a new structure for the method save_paper_mc_server() was developed. The structure contains both, executing of the applications and handling PaperMcServer calls. This change ensures a clearer data flow and increases code readability.

This commit also introduces a new macro required_def!(), to avoid code repetition when accessing definitions that must be provided in the PaperMcServer request payload."